navigator is now a shell script that hides the default map file to use. (I refuse to hardcode it in the executable)
The actual executable that is compiled is called navigate.

Added some more cmake functionality to copy over some directories from the source tree if doing an out of source build. (maps and points)

Maps are now loaded from the ./maps directory

Most of the shell for the navigation executable is finished. Compiles without error but does nothing past setting the map and sensors up.